Refactor: login input hook custom
