+++ src/pages/_app.tsx
@@ -1,12 +1,17 @@
+// _app.tsx 파일
 import React from "react";
 import { AppProps } from "next/app";
-import { StoreProvider } from "@/store/StoreProvider";
-import "@/styles/globals.css";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { StoreProvider } from "@/store/StoreProvider";
+import { useRouter } from "next/router";
+
+import "@/styles/globals.css";
 
 const queryClient = new QueryClient();
 
 const MyApp: React.FC<AppProps> = ({ Component, pageProps }) => {
+  const router = useRouter();
+
   return (
     <QueryClientProvider client={queryClient}>
       <StoreProvider>
